% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cache_sfeLung}
\alias{cache_sfeLung}
\title{cache and/or retrieve path to an SFE of V1 lung demo data from 10x}
\usage{
cache_sfeLung(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/lungSFEtxg.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where zip file can be retrieved}
}
\description{
cache and/or retrieve path to an SFE of V1 lung demo data from 10x
}
\note{
We are explicitly avoiding declaring reliance on terra
or SpatialFeatureExperiment, to keep package weight low.
Thus the example will fail unless these are already present.
}
\examples{
if (!requireNamespace("terra")) {
  message("install terra package to run this example")
  } else if (!requireNamespace("SpatialFeatureExperiment")) {
  message("install SpatialFeatureExperiment package to run this example")
  } else {
  zp = cache_sfeLung()
  td = tempdir()
  unzip(zp, exdir=td)
  ans = HDF5Array::loadHDF5SummarizedExperiment(file.path(td, "lungSFEtxg"))
  SpatialFeatureExperiment::show(ans)
  }
}
