% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cache_sfeLung_ntx}
\alias{cache_sfeLung_ntx}
\title{cache and/or retrieve path to an SFE of V1 lung demo data from 10x}
\usage{
cache_sfeLung_ntx(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/sfeLung.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where zip file can be retrieved}
}
\description{
cache and/or retrieve path to an SFE of V1 lung demo data from 10x
}
\note{
Lacks transcript coordinates
}
