% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cache_xen_luad}
\alias{cache_xen_luad}
\title{cache and/or retrieve path to Xenium Lung Adenocarcinoma example data, zipped SPEP
accompanied by parquet}
\usage{
cache_xen_luad(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/luad2.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where tiff file can be retrieved}
}
\description{
cache and/or retrieve path to Xenium Lung Adenocarcinoma example data, zipped SPEP
accompanied by parquet
}
\examples{
if (interactive()) {
   pa = cache_xen_luad()
   luad = restoreZipXenSPEP(pa)
   print(luad)
   print(slot(luad, "cellbounds_path"))
   view_seg(luad, xlim=c(4000,4500), ylim=c(2000,2500))
}
}
