% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cache_xen_pdmel_lite}
\alias{cache_xen_pdmel_lite}
\title{counts-in-memory version of melanoma 5k dataset}
\usage{
cache_xen_pdmel_lite(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/pdmel_lite.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where zip file can be retrieved}
}
\description{
counts-in-memory version of melanoma 5k dataset
}
\examples{
if (interactive()) {
   pa = cache_xen_pdmel_lite()
   pdmel_lite = restoreZipXenSPEP(pa)
   print(pdmel_lite)
}
}
