% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cache_xen_sk}
\alias{cache_xen_sk}
\title{cache and/or retrieve path to xenium human primary melanoma, 5k genes}
\usage{
cache_xen_sk(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/xen_sk_spep.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where tiff file can be retrieved}
}
\description{
cache and/or retrieve path to xenium human primary melanoma, 5k genes
}
\note{
Retrieves path to a 1.05GB zip file.  The zip file
includes a 'saved' HDF5-backed SpatialExperiment and
parquet files for cell, nucleus, and transcript coordinates.
The example shows how these can be assembled into a
XenSPEP instance.
}
\examples{
if (interactive()) {
 td = tempdir()
 pzip_path = cache_xen_sk()
 dir.create(xpw <- file.path(td, "xen_sk_work"))
 unzip(pzip_path, exdir=xpw)
 mel = HDF5Array::loadHDF5SummarizedExperiment(file.path(xpw, "xen_sk"))
 mel = reset_parq_paths(mel, xpw)
 mel = loadGeometry(mel)
 print(slot(mel, "cbtab") |> head() |> dplyr::collect())
 print(slot(mel, "txtab") |> dplyr::count() |> dplyr::collect())
 plot(SpatialExperiment::spatialCoords(mel), pch=".",
   main="10x FFPE Human dermal melanoma sample")
}
}
