% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_10x_test_data.R
\name{cacheXenBr2fov}
\alias{cacheXenBr2fov}
\title{acquire path to cached 2 field of view test data from Xenium 2.0.0 test data repo, breast tissue}
\usage{
cacheXenBr2fov(
  cache = BiocFileCache::BiocFileCache(),
  url =
    "https://cf.10xgenomics.com/samples/xenium/2.0.0/Xenium_V1_human_Breast_2fov/Xenium_V1_human_Breast_2fov_outs.zip"
)
}
\arguments{
\item{cache}{like BiocFileCache}

\item{url}{character(1) defaults to value valid at 27 Nov 2024}
}
\description{
acquire path to cached 2 field of view test data from Xenium 2.0.0 test data repo, breast tissue
}
\examples{
br2pa = cacheXenBr2fov()
td = tempfile()
unzip(br2pa, exdir=td)
br2fov = ingest_xen(td)
br2fov = loadGeometry(br2fov)
br2fov
br2fov@cbtab |> as.data.frame() |> lapply(range)
viewSeg(br2fov, c(0,1231), c(1, 748))
}
