% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_mtif.R
\name{cacheXenProstLite}
\alias{cacheXenProstLite}
\title{counts-in-memory version of prostate 5k dataset}
\usage{
cacheXenProstLite(
  cache = BiocFileCache::BiocFileCache(),
  url = "https://mghp.osn.xsede.org/bir190004-bucket01/BiocXenData/prost_lite.zip"
)
}
\arguments{
\item{cache}{defaults to BiocFileCache::BiocFileCache()}

\item{url}{location where zip file can be retrieved}
}
\value{
path to cached resource
}
\description{
counts-in-memory version of prostate 5k dataset
}
\examples{
if (interactive()) {
   pa = cacheXenProstLite()
   prost_lite = restoreZipXenSPEP(pa)
   print(prost_lite)
}
}
