% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellBoundaries.R
\name{clipRect}
\alias{clipRect}
\title{restrict XenSPEP to cells with centroids in specified rectangle,
also restrict boundary and transcript location features}
\usage{
clipRect(xsce, xlim, ylim)
}
\arguments{
\item{xsce}{XenSPEP instance}

\item{xlim}{numeric(2)}

\item{ylim}{numeric(2)}
}
\value{
XenSPEP instance
}
\description{
restrict XenSPEP to cells with centroids in specified rectangle,
also restrict boundary and transcript location features
}
\note{
Could be too RAM-hungry.
}
\examples{
args(clipRect)
}
