% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebapp.R
\name{demoapp}
\alias{demoapp}
\title{simple app to explore an image}
\usage{
demoapp(simple = FALSE)
}
\arguments{
\item{simple}{logical(1) if TRUE, use a cached tiff for illustration}
}
\description{
simple app to explore an image
}
\note{
Navigate file input control to location of tiffs
}
\examples{
if (interactive()) demoapp(simple=TRUE)
}
