% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e2sym.R
\name{e2sym}
\alias{e2sym}
\title{helper function to map ENS ids to symbols}
\usage{
e2sym(x)
}
\arguments{
\item{x}{character() mix of Ensembl Ids and other strings; the latter are left unchanged}
}
\value{
a vector like x with gene symbols from v79 mapping substituted where possible
}
\description{
helper function to map ENS ids to symbols
}
\examples{
e2sym(c("ABC", "ENSG00000213088", "ENSG00000107796", "ENSG00000163017"))
}
