% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e79sym}
\alias{e79sym}
\title{mapping from ENSG to symbols based on EnsDb.Hsapiens.v79}
\format{
character()
}
\usage{
e79sym
}
\description{
mapping from ENSG to symbols based on EnsDb.Hsapiens.v79
}
\note{
named character vector with values gene symbols, name ENSG ids
}
\examples{
data(e79sym)
head(e79sym)
}
\keyword{datasets}
