% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getCellBoundaries,XenSPEP-method}
\alias{getCellBoundaries,XenSPEP-method}
\title{method for cell boundary extraction}
\usage{
\S4method{getCellBoundaries}{XenSPEP}(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
method for cell boundary extraction
}
