% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getNucleusBoundaries,XenSPEP-method}
\alias{getNucleusBoundaries,XenSPEP-method}
\title{method for nucleus boundary extraction}
\usage{
\S4method{getNucleusBoundaries}{XenSPEP}(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
method for nucleus boundary extraction
}
