% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getNucleusBoundaries}
\alias{getNucleusBoundaries}
\title{method for nucleus boundary extraction}
\usage{
getNucleusBoundaries(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
method for nucleus boundary extraction
}
\examples{
showMethods("getNucleusBoundaries")
}
