% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getTranscripts,XenSPEP-method}
\alias{getTranscripts,XenSPEP-method}
\title{method for transcript extraction}
\usage{
\S4method{getTranscripts}{XenSPEP}(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
method for transcript extraction
}
