% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getTranscripts}
\alias{getTranscripts}
\title{method for transcript extraction}
\usage{
getTranscripts(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\value{
reference to ingested parquet
}
\description{
method for transcript extraction
}
\examples{
showMethods("getTranscripts")
}
