% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGeometry-XenSPEP.R
\name{loadGeometry,XenSPEP-method}
\alias{loadGeometry,XenSPEP-method}
\title{read and bind parquet data to XenSPEP}
\usage{
\S4method{loadGeometry}{XenSPEP}(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\value{
instance of XenSPEP
}
\description{
read and bind parquet data to XenSPEP
}
