% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGeometry-XenSPEP.R
\name{loadGeometry}
\alias{loadGeometry}
\title{read and bind parquet data to XenSPEP}
\usage{
loadGeometry(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
read and bind parquet data to XenSPEP
}
