% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_zarr.R
\name{make_cropper}
\alias{make_cropper}
\title{use spatialdata bounding_box_query to confine attention}
\usage{
make_cropper(xlim = c(19000, 22000), ylim = c(8000, 8500))
}
\arguments{
\item{xlim}{extent of interest, 'horizontal' axis}

\item{ylim}{extent of interest, 'vertical' axis}
}
\value{
py_run_string output, which is a python environment,
which will have element 'crop0' to use as python function
}
\description{
use spatialdata bounding_box_query to confine attention
}
