% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcode.R
\name{plot_xen_ggprep}
\alias{plot_xen_ggprep}
\title{plot method for ggplot2-prepared XenSPEP}
\usage{
plot_xen_ggprep(x, y, ...)
}
\arguments{
\item{x}{instance of S3 class `xen_ggprep`}

\item{y}{not used}

\item{\dots}{not used}
}
\description{
plot method for ggplot2-prepared XenSPEP
}
\note{
roxygen had problems with this
}
