% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcode.R
\name{plotXenGgprep}
\alias{plotXenGgprep}
\title{plot method for ggplot2-prepared XenSPEP}
\usage{
plotXenGgprep(x, y, ...)
}
\arguments{
\item{x}{instance of S3 class `xen_ggprep`}

\item{y}{not used}

\item{\dots}{not used}
}
\value{
ggplot
}
\description{
plot method for ggplot2-prepared XenSPEP
}
\note{
roxygen had problems with this
}
\examples{
pa = cacheXenLuad()
luad = restoreZipXenSPEP(pa)
hh = ggprepSeg(luad, c(4000,4500), c(2000,2500))
plotXenGgprep(hh)
}
