% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcode.R
\name{print_xen_ggprep}
\alias{print_xen_ggprep}
\title{print method for ggplot2-prepared XenSPEP}
\usage{
print_xen_ggprep(x, ...)
}
\arguments{
\item{x}{instance of S3 class `xen_ggprep`}

\item{\dots}{not used}
}
\description{
print method for ggplot2-prepared XenSPEP
}
