% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rast_zarr.R
\name{rast_zarr}
\alias{rast_zarr}
\title{produce an array from selected image in a spatialdata zarr store}
\usage{
rast_zarr(
  zpath,
  xlim = c(19000, 22000),
  ylim = c(8000, 8500),
  image_name = "he_image",
  target_width = 800
)
}
\arguments{
\item{zpath}{path to zarr store}

\item{xlim}{extent of interest, 'horizontal' axis}

\item{ylim}{extent of interest, 'vertical' axis}

\item{image_name}{character(1)}

\item{target_width}{numeric(1)}
}
\value{
R array dimensions nx x ny x 3 suitable for use
with EBImage
}
\description{
produce an array from selected image in a spatialdata zarr store
}
\examples{
\donttest{
if (dir.exists("lungdata.zarr")) {
  nn = rast_zarr("lungdata.zarr")
  if (requireNamespace("EBImage")) {
     ii = EBImage::Image(nn, colormode="Color")
     EBImage::display(ii) # to browser for zoom/pan
  } else message("Install EBImage to run this code.")
}
}
}
