% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_parq_paths.R
\name{resetParqPaths}
\alias{resetParqPaths}
\title{utility for dealing with cached Xen_SPEP in temp folder}
\usage{
resetParqPaths(xsp, base)
}
\arguments{
\item{xsp}{instance of XenSPEP}

\item{base}{folder path where parquet files are found}
}
\value{
XenSPEP instance
}
\description{
utility for dealing with cached Xen_SPEP in temp folder
}
\note{
Will prepend current folder path to parquet-oriented
slot values.
}
