% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipXenSPEP.R
\name{restoreZipXenSPEP}
\alias{restoreZipXenSPEP}
\title{use unzip, readRDS, and loadGeometry to restore a XenSPEP}
\usage{
restoreZipXenSPEP(zipf, exdir = tempdir())
}
\arguments{
\item{zipf}{character(1) path to zip file created with `zipXenSPEP`}

\item{exdir}{character(1) defaults to tempdir(), where contents are unpacked}
}
\value{
instance of XenSPEP
}
\description{
use unzip, readRDS, and loadGeometry to restore a XenSPEP
}
\note{
Session folder position will change with setwd(), on.exit ensures return to 
position when started.
}
