% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{show,XenSPEP-method}
\alias{show,XenSPEP-method}
\title{display aspects of XenSPEP}
\usage{
\S4method{show}{XenSPEP}(object)
}
\arguments{
\item{object}{instance of XenSPEP}
}
\value{
operates with cat()
}
\description{
display aspects of XenSPEP
}
