% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{[,XenSPEP,ANY,ANY,ANY-method}
\alias{[,XenSPEP,ANY,ANY,ANY-method}
\title{formal bracket definition, that leaves parquet geometry information alone.}
\usage{
\S4method{[}{XenSPEP,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{instance of XenSPEP}

\item{i}{feature selection}

\item{j}{cell selection}

\item{\dots}{passed to SpatialExperiment methods}

\item{drop}{logical(1)}
}
\value{
XenSPEP instance
}
\description{
formal bracket definition, that leaves parquet geometry information alone.
}
\note{
Gives a message and calls callNextMethod.
}
