% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_seg.R
\name{view_seg}
\alias{view_seg}
\title{naive polygon viewer}
\usage{
view_seg(x, xlim, ylim, show_tx = FALSE, ...)
}
\arguments{
\item{x}{instance of XenSPEP}

\item{xlim}{numeric(2) ordered vector of max and min on x}

\item{ylim}{numeric(2) ordered vector of max and min on y}

\item{show_tx}{logical(1) display transcript locations if TRUE, defaults to FALSE.}

\item{\dots}{passed to polygon()}
}
\description{
naive polygon viewer
}
\note{
This is more RAM-sparing than clip_rect followed by view.
}
\examples{
luad = cache_xen_luad()
pa = cache_xen_luad()
luad = restoreZipXenSPEP(pa)
rownames(luad) = make.names(SummarizedExperiment:::rowData(luad)$Symbol, unique=TRUE)
out = view_seg(luad, c(5800, 6300), c(1300, 1800), lwd=.5)
out$ncells
}
