% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{xdims}
\alias{xdims}
\title{helper function for XenSCE show method, producing dimensions for
geometry information}
\usage{
xdims(x)
}
\arguments{
\item{x}{instance of XenSCE}
}
\description{
helper function for XenSCE show method, producing dimensions for
geometry information
}
