% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-constructor.R
\name{XenSPEP}
\alias{XenSPEP}
\title{XenSPEP (SpatialExperiment with parquet references) constructor}
\usage{
XenSPEP(folder)
}
\arguments{
\item{folder}{character(1) 'standard' Xenium output folder}
}
\description{
XenSPEP (SpatialExperiment with parquet references) constructor
}
