% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipXenSPEP.R
\name{zipXenSPEP}
\alias{zipXenSPEP}
\title{serialize the collection of XenSPEP and parquet with zip}
\usage{
zipXenSPEP(xsp, targetfile)
}
\arguments{
\item{xsp}{instance of XenSPEP with geometry loaded}

\item{targetfile}{character(1) destination of zip process}
}
\description{
serialize the collection of XenSPEP and parquet with zip
}
\note{
a .rds and three parquet files are zipped together for restoration
by `restoreZipXenSPEP`.  The outcome is `paste0(targetfile, ".zip")`.
}
