% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenSPEP-class.R
\name{getCellBoundaries}
\alias{getCellBoundaries}
\title{method for cell boundary extraction}
\usage{
getCellBoundaries(x)
}
\arguments{
\item{x}{instance of XenSPEP}
}
\description{
method for cell boundary extraction
}
