% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcode.R
\name{ggprep_seg}
\alias{ggprep_seg}
\title{prepare a XenSPEP for ggplot2 visualization}
\usage{
ggprep_seg(xsce, xlim = c(5800, 6200), ylim = c(6300, 6700))
}
\arguments{
\item{xsce}{XenSPEP instance}

\item{xlim}{numeric(2)}

\item{ylim}{numeric(2)}
}
\value{
a list with components `bounds` (data.frame including relevant colData
rows (all colData variables) and cell boundary coordinates) and `txdata`, a filtered
arrow Table.
}
\description{
prepare a XenSPEP for ggplot2 visualization
}
\note{
This is idiosyncratic.  Quintiles of cell_area are produced, and transcript locations
are filtered
}
