% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_xen.R
\name{ingest_xen}
\alias{ingest_xen}
\title{produce a pre-loaded XenSPEP (SpatialExperiment with parquet references)}
\usage{
ingest_xen(folder)
}
\arguments{
\item{folder}{character(1) 'standard' Xenium output folder}
}
\description{
produce a pre-loaded XenSPEP (SpatialExperiment with parquet references)
}
