% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_parq_paths.R
\name{reset_parq_paths}
\alias{reset_parq_paths}
\title{utility for dealing with cached Xen_SPEP in temp folder}
\usage{
reset_parq_paths(xsp, base)
}
\arguments{
\item{xsp}{instance of XenSPEP}

\item{base}{folder path where parquet files are found}
}
\description{
utility for dealing with cached Xen_SPEP in temp folder
}
\note{
Will prepend current folder path to parquet-oriented
slot values.
}
\examples{
prp = cache_xen_prost() # parquet paths are pre-set :(
dir.create(td <- file.path(tempdir(), "xen_work"))
unzip(prp, exdir=td)
ini = HDF5Array::loadHDF5SummarizedExperiment(file.path(td, "xen_prost"))
fin = reset_parq_paths(ini, base=td)
print(fin)
txloc = arrow::read_parquet(slot(fin, "tx_path"), as_data_frame=FALSE) # premature
txloc |> dplyr::count() |> dplyr::collect()
fin = loadGeometry(fin) # proper
slot(fin, "txtab")
slot(fin, "txtab") |> dplyr::count() |> dplyr::collect()
}
